add security

+++ src/items/items.router.ts
@@ -5,6 +5,9 @@
  import * as ItemService from "./items.service";
  import { BaseItem, Item } from "./item.interface";
 
+ import { checkJwt } from "../middleware/authz.middleware";
+ import { checkPermissions } from "../middleware/permissions.middleware";
+ import { ItemPermission } from "./item-permission";
 /**
  * Router Definition
  */
@@ -42,8 +45,14 @@ itemsRouter.get("/:id", async (req: Request, res: Response) => {
     }
   });
 
+// SECURE PATHS BELOW
+itemsRouter.use(checkJwt); 
+
 // POST items
-itemsRouter.post("/", async (req: Request, res: Response) => {
+itemsRouter.post(
+  "/",
+  checkPermissions(ItemPermission.CreateItems),
+   async (req: Request, res: Response) => {
     try {
       const item: BaseItem = req.body;
   
@@ -57,7 +66,10 @@ itemsRouter.post("/", async (req: Request, res: Response) => {
   
 
 // PUT items/:id
-itemsRouter.put("/:id", async (req: Request, res: Response) => {
+itemsRouter.put(
+  "/:id",
+  checkPermissions(ItemPermission.UpdateItems),
+  async (req: Request, res: Response) => {
     const id: number = parseInt(req.params.id, 10);
   
     try {
@@ -79,7 +91,10 @@ itemsRouter.put("/:id", async (req: Request, res: Response) => {
   });
 
 // DELETE items/:id
-itemsRouter.delete("/:id", async (req: Request, res: Response) => {
+itemsRouter.delete(
+  "/:id",
+  checkPermissions(ItemPermission.DeleteItems),
+  async (req: Request, res: Response) => {
     try {
       const id: number = parseInt(req.params.id, 10);
       await ItemService.remove(id);
